requirements.txt
streamlit
pandas
gspread
google-generativeai
pdfplumber